Add your annotation as a separate page
